+++ tests/crawlers/test_cases/twitch.py
@@ -4,8 +4,8 @@ TEST_CASES = [
         "url": "https://www.twitch.tv/fanfan/clip/GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM",
         "results": [
             {
-                "url": "re:expires=",
-                "filename": "not scared btw [GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM][60fps][1080p].mp4",
+                "url": "re:/index.mp4?token=",
+                "filename": "not scared btw [GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM][1080p].mp4",
                 "debrid_url": None,
                 "original_filename": "not scared btw",
                 "referer": "https://clips.twitch.tv/GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM",
@@ -19,8 +19,8 @@ TEST_CASES = [
         "url": "https://clips.twitch.tv/GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM",
         "results": [
             {
-                "url": "re:expires=",
-                "filename": "not scared btw [GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM][60fps][1080p].mp4",
+                "url": "re:/index.mp4?token=",
+                "filename": "not scared btw [GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM][1080p].mp4",
                 "referer": "https://clips.twitch.tv/GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM",
             }
         ],
@@ -29,8 +29,8 @@ TEST_CASES = [
         "url": "https://clips.twitch.tv/embed?clip=GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM",
         "results": [
             {
-                "url": "re:/h0073-YuUM9o4kAK7QUTjA/AT-cm%7Ch0073-YuUM9o4kAK7QUTjA.mp4",
-                "filename": "not scared btw [GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM][60fps][1080p].mp4",
+                "url": "re:/index.mp4?token=",
+                "filename": "not scared btw [GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM][1080p].mp4",
                 "referer": "https://clips.twitch.tv/GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM",
             }
         ],
@@ -40,7 +40,7 @@ TEST_CASES = [
         "results": [
             {
                 "url": "ANY",
-                "filename": "not scared btw [GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM][60fps][1080p].mp4",
+                "filename": "not scared btw [GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM][1080p].mp4",
                 "referer": "https://clips.twitch.tv/GeniusDifficultSproutPhilosoraptor-5i5Qz2jiQArtBROM",
             }
         ],
@@ -57,21 +57,6 @@ TEST_CASES = [
                 "referer": "https://www.twitch.tv/videos/1744169736",
                 "album_id": None,
                 "uploaded_at": 1676928416,
-                "download_folder": "re:Loose Files (Twitch)",
-            }
-        ],
-    },
-    {
-        "url": "https://clips.twitch.tv/LitigiousNaiveReubenUWot-E0uuswUYOhs9htgV",
-        "results": [
-            {
-                "url": "ANY",
-                "filename": "IM GUNNA MAKE IT... [LitigiousNaiveReubenUWot-E0uuswUYOhs9htgV][60fps][1440p].mp4",
-                "debrid_url": None,
-                "original_filename": "IM GUNNA MAKE IT...",
-                "referer": "https://clips.twitch.tv/LitigiousNaiveReubenUWot-E0uuswUYOhs9htgV",
-                "album_id": None,
-                "uploaded_at": 1767765454,
                 "download_folder": "re:Loose Files (Twitch)",
             }
         ],